# bpo-26415: reduce peak memory consumption by the parser