--- a/.gitattributes
+++ b/.gitattributes
@@ -59,3 +59,4 @@ Doc/library/token-list.inc  linguist-generated=true
 Include/token.h             linguist-generated=true
 Lib/token.py                linguist-generated=true
 Parser/token.c              linguist-generated=true
+Modules/gramvalid.c         linguist-generated=true