--- a/Makefile.pre.in
+++ b/Makefile.pre.in
@@ -794,9 +794,11 @@ regen-grammar: regen-token
 	$(PYTHON_FOR_REGEN) -m Parser.pgen $(srcdir)/Grammar/Grammar \
 		$(srcdir)/Grammar/Tokens \
 		$(srcdir)/Include/graminit.h.new \
-		$(srcdir)/Python/graminit.c.new
+		$(srcdir)/Python/graminit.c.new \
+		$(srcdir)/Modules/gramvalid.c.new
 	$(UPDATE_FILE) $(srcdir)/Include/graminit.h $(srcdir)/Include/graminit.h.new
 	$(UPDATE_FILE) $(srcdir)/Python/graminit.c $(srcdir)/Python/graminit.c.new
+	$(UPDATE_FILE) $(srcdir)/Modules/gramvalid.c $(srcdir)/Modules/gramvalid.c.new
 
 .PHONY=regen-ast
 regen-ast: