--- a/Misc/NEWS.d/next/Core and Builtins/2018-12-06-16-55-19.bpo-26415.jbufv5.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2018-12-06-16-55-19.bpo-26415.jbufv5.rst
@@ -1,0 +1,3 @@
+Reduce peak memory consumption by the parser, by compressing some
+branchless stretches in the parse tree into a single node. AST is
+not affected.