@tyomitch 2018-12-12
When I leave my machine alone while it's running the benchmarks, the results are even more exciting: all significant differences are in favour of the patched parser, ranging from "1.03x faster" to "1.26x faster".
```
### 2to3 ###
Mean +- std dev: 709 ms +- 62 ms -> 674 ms +- 59 ms: 1.05x faster
Significant (t=3.17)

### chaos ###
Mean +- std dev: 308 ms +- 36 ms -> 292 ms +- 39 ms: 1.06x faster
Significant (t=2.45)

### crypto_pyaes ###
Mean +- std dev: 318 ms +- 31 ms -> 280 ms +- 43 ms: 1.14x faster
Significant (t=5.55)

### deltablue ###
Mean +- std dev: 18.5 ms +- 2.0 ms -> 16.3 ms +- 2.5 ms: 1.13x faster
Significant (t=5.13)

### django_template ###
Mean +- std dev: 286 ms +- 16 ms -> 265 ms +- 25 ms: 1.08x faster
Significant (t=5.34)

### dulwich_log ###
Mean +- std dev: 133 ms +- 11 ms -> 131 ms +- 10 ms: 1.01x faster
Not significant

### fannkuch ###
Mean +- std dev: 1.52 sec +- 0.08 sec -> 1.36 sec +- 0.11 sec: 1.12x faster
Significant (t=8.86)

### float ###
Mean +- std dev: 353 ms +- 55 ms -> 298 ms +- 72 ms: 1.18x faster
Significant (t=4.72)

### go ###
Mean +- std dev: 658 ms +- 48 ms -> 562 ms +- 69 ms: 1.17x faster
Significant (t=8.86)

### hexiom ###
Mean +- std dev: 26.8 ms +- 6.8 ms -> 24.9 ms +- 5.1 ms: 1.08x faster
Not significant

### html5lib ###
Mean +- std dev: 223 ms +- 28 ms -> 198 ms +- 29 ms: 1.12x faster
Significant (t=4.74)

### json_dumps ###
Mean +- std dev: 34.3 ms +- 4.3 ms -> 31.9 ms +- 4.7 ms: 1.08x faster
Significant (t=2.98)

### json_loads ###
Mean +- std dev: 59.8 us +- 4.4 us -> 56.8 us +- 7.4 us: 1.05x faster
Significant (t=2.74)

### logging_format ###
Mean +- std dev: 22.4 us +- 1.6 us -> 21.3 us +- 2.6 us: 1.05x faster
Significant (t=2.82)

### logging_silent ###
Mean +- std dev: 511 ns +- 70 ns -> 475 ns +- 79 ns: 1.07x faster
Significant (t=2.60)

### logging_simple ###
Mean +- std dev: 20.9 us +- 1.4 us -> 19.7 us +- 1.9 us: 1.06x faster
Significant (t=3.64)

### mako ###
Mean +- std dev: 47.9 ms +- 6.7 ms -> 41.2 ms +- 8.0 ms: 1.16x faster
Significant (t=4.95)

### meteor_contest ###
Mean +- std dev: 247 ms +- 31 ms -> 227 ms +- 33 ms: 1.08x faster
Significant (t=3.24)

### nbody ###
Mean +- std dev: 464 ms +- 65 ms -> 370 ms +- 82 ms: 1.26x faster
Significant (t=7.01)

### nqueens ###
Mean +- std dev: 307 ms +- 28 ms -> 266 ms +- 48 ms: 1.16x faster
Significant (t=5.82)

### pathlib ###
Mean +- std dev: 47.4 ms +- 5.2 ms -> 44.4 ms +- 5.9 ms: 1.07x faster
Significant (t=2.89)

### pickle ###
Mean +- std dev: 18.8 us +- 2.3 us -> 17.6 us +- 2.4 us: 1.07x faster
Significant (t=2.77)

### pickle_dict ###
Mean +- std dev: 62.9 us +- 4.9 us -> 58.3 us +- 7.6 us: 1.08x faster
Significant (t=3.97)

### pickle_list ###
Mean +- std dev: 8.65 us +- 0.53 us -> 8.22 us +- 0.77 us: 1.05x faster
Significant (t=3.59)

### pickle_pure_python ###
Mean +- std dev: 1.22 ms +- 0.14 ms -> 1.12 ms +- 0.16 ms: 1.08x faster
Significant (t=3.51)

### pidigits ###
Mean +- std dev: 202 ms +- 4 ms -> 205 ms +- 9 ms: 1.02x slower
Not significant

### python_startup ###
Mean +- std dev: 12.5 ms +- 1.1 ms -> 12.2 ms +- 1.1 ms: 1.02x faster
Not significant

### python_startup_no_site ###
Mean +- std dev: 8.74 ms +- 0.79 ms -> 8.86 ms +- 0.70 ms: 1.01x slower
Not significant

### raytrace ###
Mean +- std dev: 1.30 sec +- 0.08 sec -> 1.18 sec +- 0.11 sec: 1.11x faster
Significant (t=7.10)

### regex_compile ###
Mean +- std dev: 449 ms +- 42 ms -> 427 ms +- 46 ms: 1.05x faster
Significant (t=2.72)

### regex_dna ###
Mean +- std dev: 546 ms +- 43 ms -> 520 ms +- 51 ms: 1.05x faster
Significant (t=3.03)

### regex_effbot ###
Mean +- std dev: 12.1 ms +- 1.2 ms -> 11.1 ms +- 1.4 ms: 1.09x faster
Significant (t=4.10)

### regex_v8 ###
Mean +- std dev: 77.9 ms +- 6.4 ms -> 74.5 ms +- 8.6 ms: 1.05x faster
Significant (t=2.50)

### richards ###
Mean +- std dev: 179 ms +- 20 ms -> 169 ms +- 26 ms: 1.06x faster
Significant (t=2.42)

### scimark_fft ###
Mean +- std dev: 1.32 sec +- 0.09 sec -> 1.21 sec +- 0.11 sec: 1.10x faster
Significant (t=6.25)

### scimark_lu ###
Mean +- std dev: 519 ms +- 55 ms -> 485 ms +- 55 ms: 1.07x faster
Significant (t=3.34)

### scimark_monte_carlo ###
Mean +- std dev: 339 ms +- 44 ms -> 305 ms +- 47 ms: 1.11x faster
Significant (t=4.11)

### scimark_sor ###
Mean +- std dev: 643 ms +- 39 ms -> 577 ms +- 58 ms: 1.11x faster
Significant (t=7.27)

### scimark_sparse_mat_mult ###
Mean +- std dev: 20.9 ms +- 2.7 ms -> 18.9 ms +- 4.1 ms: 1.10x faster
Significant (t=3.10)

### spectral_norm ###
Mean +- std dev: 484 ms +- 52 ms -> 433 ms +- 64 ms: 1.12x faster
Significant (t=4.80)

### sqlalchemy_declarative ###
Mean +- std dev: 297 ms +- 23 ms -> 291 ms +- 28 ms: 1.02x faster
Not significant

### sqlalchemy_imperative ###
Mean +- std dev: 47.4 ms +- 3.6 ms -> 45.8 ms +- 4.3 ms: 1.03x faster
Significant (t=2.18)

### sqlite_synth ###
Mean +- std dev: 9.07 us +- 1.43 us -> 8.52 us +- 1.16 us: 1.06x faster
Significant (t=2.31)

### sympy_expand ###
Mean +- std dev: 923 ms +- 62 ms -> 859 ms +- 71 ms: 1.07x faster
Significant (t=5.28)

### sympy_integrate ###
Mean +- std dev: 40.8 ms +- 4.2 ms -> 38.3 ms +- 5.2 ms: 1.07x faster
Significant (t=2.95)

### sympy_str ###
Mean +- std dev: 430 ms +- 35 ms -> 394 ms +- 47 ms: 1.09x faster
Significant (t=4.77)

### sympy_sum ###
Mean +- std dev: 195 ms +- 12 ms -> 180 ms +- 19 ms: 1.09x faster
Significant (t=5.42)

### telco ###
Mean +- std dev: 17.4 ms +- 1.8 ms -> 16.6 ms +- 1.6 ms: 1.05x faster
Significant (t=2.61)

### tornado_http ###
Mean +- std dev: 319 ms +- 14 ms -> 310 ms +- 21 ms: 1.03x faster
Significant (t=2.63)

### unpack_sequence ###
Mean +- std dev: 216 ns +- 46 ns -> 185 ns +- 50 ns: 1.17x faster
Significant (t=3.57)

### unpickle ###
Mean +- std dev: 31.9 us +- 4.0 us -> 31.3 us +- 4.6 us: 1.02x faster
Not significant

### unpickle_list ###
Mean +- std dev: 9.72 us +- 0.94 us -> 9.33 us +- 0.98 us: 1.04x faster
Significant (t=2.22)

### unpickle_pure_python ###
Mean +- std dev: 1.01 ms +- 0.13 ms -> 0.91 ms +- 0.14 ms: 1.11x faster
Significant (t=4.07)

### xml_etree_generate ###
Mean +- std dev: 293 ms +- 32 ms -> 278 ms +- 34 ms: 1.05x faster
Significant (t=2.37)

### xml_etree_iterparse ###
Mean +- std dev: 312 ms +- 29 ms -> 297 ms +- 37 ms: 1.05x faster
Significant (t=2.44)

### xml_etree_parse ###
Mean +- std dev: 445 ms +- 40 ms -> 430 ms +- 43 ms: 1.03x faster
Not significant

### xml_etree_process ###
Mean +- std dev: 237 ms +- 18 ms -> 224 ms +- 27 ms: 1.06x faster
Significant (t=3.29)
```