@methane 2019-04-05
I can not believe this affects runtime performance.  I tested this PR on my machine (dedicated Linux box).

```
$ pyperf compare_to master.json 10995.json -G --min-speed=1
Slower (2):
- pickle_dict: 53.2 us +- 0.2 us -> 54.0 us +- 2.2 us: 1.02x slower (+2%)
- scimark_monte_carlo: 262 ms +- 2 ms -> 265 ms +- 4 ms: 1.01x slower (+1%)

Faster (1):
- unpickle: 30.2 us +- 2.3 us -> 28.7 us +- 2.2 us: 1.05x faster (-5%)

Benchmark hidden because not significant (54):
```